update : MyProfile destructuring in RepresentativeProfile and DancerProfile
